Extract window handling to separate file
